자신의 댓글만 삭제 가능

+++ src/main/java/com/prj2spring20240521/controller/comment/CommentController.java
@@ -3,6 +3,7 @@ package com.prj2spring20240521.controller.comment;
 import com.prj2spring20240521.domain.comment.Comment;
 import com.prj2spring20240521.service.comment.CommentService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
@@ -33,8 +34,11 @@ public class CommentController {
     }
 
     @DeleteMapping("remove")
-    public void remove(@RequestBody Comment comment) {
-        System.out.println("comment = " + comment);
-        service.remove(comment);
+    public ResponseEntity remove(@RequestBody Comment comment, Authentication authentication) {
+        if (service.hasAccess(comment, authentication)) {
+            service.remove(comment);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 }
